Remove cacheprovider wrapper and accessories  into CacheBasedOperationHolder.
