+++ Src/DependencyCollector/Shared/Implementation/DependencyTableStore.cs
@@ -17,8 +17,8 @@
         private DependencyTableStore() 
         {
 #if !NET40
-            this.WebRequestCacheHolder = new CacheBasedOperationHolder();
-            this.SqlRequestCacheHolder = new CacheBasedOperationHolder();
+            this.WebRequestCacheHolder = new CacheBasedOperationHolder("aisdkwebrequests", 100 * 1000);
+            this.SqlRequestCacheHolder = new CacheBasedOperationHolder("aisdksqlrequests", 100 * 1000);
 #endif
             this.WebRequestConditionalHolder = new ObjectInstanceBasedOperationHolder();
             this.SqlRequestConditionalHolder = new ObjectInstanceBasedOperationHolder();
